refactor: Bullet related extensions method moved to Bullet CameraComponentExtensions
